Reorganize code for legibility
