+++ lib/events.js
@@ -1,0 +1,26 @@
+'use strict'
+
+var once = require('once')
+
+function onExit (fn) {
+  var oneFn = once(fn)
+  process.on('beforeExit', handle('beforeExit'))
+  process.on('exit', handle('exit'))
+  process.on('uncaughtException', handle('uncaughtException', 1))
+  process.on('SIGHUP', handle('SIGHUP', 129))
+  process.on('SIGINT', handle('SIGINT', 130))
+  process.on('SIGQUIT', handle('SIGQUIT', 131))
+  process.on('SIGTERM', handle('SIGTERM', 143))
+  function handle (evt, code) {
+    onExit.passCode = function (code) {
+      if (oneFn.value) { oneFn = once(fn) }
+      oneFn(code, evt)
+    }
+    onExit.insertCode = function () {
+      if (oneFn.value) { oneFn = once(fn) }
+      oneFn(code, evt)
+    }
+    return (code === undefined) ? onExit.passCode : onExit.insertCode
+  }
+}
+module.exports.onExit = onExit
